requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
Pillow==10.2.0
stripe==7.9.0
pandas==2.1.4
